pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm[toml]"]
build-backend = "setuptools.build_meta"

[project]
name = "embryostage"
description = "Classification of embryonic stages from label-free movies"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Shalin Mehta", email = "shalin.mehta@gmail.com" },
  { name = "Dave Matus", email = "dave.matus@arcadiascience.com" },
  { name = "Keith Cheveralls", email = "keith.cheveralls@arcadiascience.com" },
]

# TODO: pin versions of all dependencies here
dependencies = [
  "iohub==0.1.0.dev4",
  "nd2",
  "torch>=2.0.0",
  "tensorboard>=2.13.0",
  "lightning>=2.0.1",
  "pytorch_model_summary",
  "numpy",
  "qtpy",
  "pandas",
  "natsort",
  "napari[all]",
  "napari-ome-zarr",
  "matplotlib",
  # note: headless version of opencv-python is needed to avoid conflict with qtpy's plugins
  # https://forum.qt.io/topic/119109/using-pyqt5-with-opencv-python-cv2-causes-error-could-not-load-qt-platform-plugin-xcb-even-though-it-was-found/7
  "opencv-python-headless",
  "scikit-image",
  "scikit-learn",
  "torchvision",
  "torchmetrics",
  "monai",
  "tensorboard",
  "tqdm",
]
dynamic = ["version"]

[project.optional-dependencies]
metrics = ["torchmetrics[detection]>=1.0.0"]
dev = [
  "black~=22.3",
  "flake8~=5.0",
  "isort~=5.10",
  "pre-commit~=2.19",
  "pylint~=2.14",
  "pytest~=7.1",
  # TODO: remove these if unused after writing tests
  "pytest-cov",
  "hypothesis",
  "profilehooks",
  "onnxruntime",
]

[project.scripts]
convert-nd2-to-zarr = "embryostage.scripts.convert_nd2_to_zarr:main"
find-embryos = "embryostage.scripts.find_embryos:main"
view-embryos = "embryostage.scripts.view_embryos:main"
encode-dynamics = "embryostage.scripts.encode_dynamics:main"
train-models = "embryostage.scripts.train_models:main"
view-embryo-classification = "embryostage.scripts.view_embryo_classification:main"
batch-classify-embryos = "embryostage.scripts.batch_classify_embryos:main"


[tool.setuptools_scm]
write_to = "src/embryostage/_version.py"

[tool.black]
src = ["embryostage"]
line-length = 95
target-version = ['py310']
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | notebooks
    | ignore
    | _version.py
  )
)
'''

[tool.isort]
profile = "black"
line_length = 95
lines_between_types = 1
default_section = "THIRDPARTY"
no_lines_before = ["STDLIB", "THIRDPARTY"]
ensure_newline_before_comments = true
